w3c-reports/12092014/efl/runned-tests/html/semantics/embedded-content/media-elements/track/track-element/cors/032-expected.txt

CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.
Actual	Expected
{
  "event": "error",
  "requests": []
}
{
  "event": "error",
  "requests": [
    {
      "cors": "http://localhost:8800",
      "cookie": "no"
    },
    {
      "cors": "http://localhost:8800",
      "cookie": "no"
    }
  ]
}
Test id: b5a1e2d0-d108-44a2-b08e-5ed9dc581ece



FAIL track CORS: Anonymous, not same-origin, with headers, redirects to not same-origin, no headers assert_object_equals: expected property "0" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
check_equal@http://localhost:8800/resources/testharness.js:384:24
check_equal@http://localhost:8800/resources/testharness.js:375:37
assert_object_equals@http://localhost:8800/resources/testharness.js:390:21
checkData@http://localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/common.js?pipe=sub:128:29
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
